% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-strict-list.R
\name{as_strict_list}
\alias{as_strict_list}
\title{Convert an existing named list into a strict one.}
\usage{
as_strict_list(obj, name = NULL)
}
\arguments{
\item{obj}{the existing named list}

\item{name}{an optional description for the list, used in error messages}
}
\description{
Convert an existing named list into a strict one.
}
