% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors-parse.R
\name{errors_parse}
\alias{errors_parse}
\title{Parse errors from vignette}
\usage{
errors_parse(path_rmd, pattern, check = TRUE)
}
\arguments{
\item{path_rmd}{Path to the .Rmd file containing the error
descriptions}

\item{pattern}{A regular expression matching each error message,
e.g., \verb{E[0-9]\{4\}}.  This should not include beginning or end of
string markers.}

\item{check}{Logical, indicating if we should check that we can
render everything we produce}
}
\value{
A list, save this within the package
}
\description{
Parse errors from vignette
}
